+++ src/parser/enrichers/monster/SwarmOfStirges/_module.mjs
@@ -1,0 +1,1 @@
+export { default as SwarmOfProboscises } from "./SwarmOfProboscises.mjs";