+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Next Up
 
 - If using Legacy Postfix when munching, 2014 monsters would not pick up legacy spells. @noclipnate
+- Swarm of Stirges parsing tweaks.
 
 # 6.1.1